feat(marketing): intro film — scene entrance dissolves out of presenter beats
New manifest field dissolveInSec: a scene's entrance cross-fades over the
previous scene's extended tail (audio timing untouched; the one FadeIn
primitive now serves cuts and scene boundaries). Applied where the film
leaves the presenter (s2, s5) — her exit dip and the next scene's bloom
read as one continuous move instead of dip-then-pop (v2 gate note).
