+++ marketing/films/intro/manifest.json
@@ -23,6 +23,7 @@
     {
       "id": "s2-two-things",
       "mode": "vo",
+      "dissolveInSec": 0.6,
       "narration": "As we jump in, you'll notice two things. First — in Stigmer, agents are infrastructure. An agent isn't a prompt buried in somebody's code. It's a declarative resource: defined in YAML, versioned, and portable. The exact same definition runs on your laptop and in the cloud. Second — execution you can trust. Every run is durable. Agents pause and wait for human approval before doing anything consequential. Every action is auditable, and budgets keep costs under control.",
       "plannedDurationSec": 33,
       "shots": "S2a/S2d numbered title cards; S2b YAML-to-card morph; S2c laptop-to-cloud; S2e approval flash (spotlit); S2f budget flash (spotlit)",
@@ -137,6 +138,7 @@
     {
       "id": "s5-scale",
       "mode": "vo",
+      "dissolveInSec": 0.6,
       "narration": "When one agent isn't enough, you orchestrate. Workflows coordinate agents and systems into repeatable pipelines — with budgets: a hard limit on what a run may spend. Put it on a schedule, and it runs without you. And everything you've seen is open source, running on a single laptop. When your team's ready, the same definitions move to Stigmer Cloud — organizations, access control, and billing built in.",
       "plannedDurationSec": 27,
       "shots": "S5a workflow canvas; S5b budget (spotlit, scrolled to center); S5c schedule; S5d laptop-to-cloud morph",
